:recycle: refactor: migrate `knex-utils` to `src/shared`
