+++ api/lib/infrastructure/repositories/target-profile-summary-for-admin-repository.js
@@ -1,7 +1,7 @@
 import { knex } from '../../../db/knex-database-connection.js';
+import { fetchPage } from '../../../src/shared/infrastructure/utils/knex-utils.js';
 import { TargetProfileSummaryForAdmin } from '../../domain/models/TargetProfileSummaryForAdmin.js';
 import { DomainTransaction } from '../DomainTransaction.js';
-import { fetchPage } from '../utils/knex-utils.js';
 
 const findPaginatedFiltered = async function ({ filter, page }) {
   const query = knex('target-profiles')
