+++ api/src/certification/session/infrastructure/repositories/jury-session-repository.js
@@ -2,7 +2,7 @@ import { knex } from '../../../../../db/knex-database-connection.js';
 import { PGSQL_FOREIGN_KEY_VIOLATION_ERROR } from '../../../../../db/pgsql-errors.js';
 import { NotFoundError } from '../../../../../lib/domain/errors.js';
 import { JurySession, statuses } from '../../../../../lib/domain/models/JurySession.js';
-import { fetchPage } from '../../../../../lib/infrastructure/utils/knex-utils.js';
+import { fetchPage } from '../../../../shared/infrastructure/utils/knex-utils.js';
 import { logger } from '../../../../shared/infrastructure/utils/logger.js';
 import { CertificationOfficer } from '../../domain/models/CertificationOfficer.js';
