+++ api/src/devcomp/infrastructure/repositories/training-repository.js
@@ -3,8 +3,8 @@ import lodash from 'lodash';
 import { knex } from '../../../../db/knex-database-connection.js';
 import { NotFoundError } from '../../../../lib/domain/errors.js';
 import { TrainingTrigger } from '../../../../lib/domain/models/index.js';
-import { fetchPage } from '../../../../lib/infrastructure/utils/knex-utils.js';
 import { DomainTransaction } from '../../../shared/domain/DomainTransaction.js';
+import { fetchPage } from '../../../shared/infrastructure/utils/knex-utils.js';
 import { Training } from '../../domain/models/Training.js';
 import { TrainingForAdmin } from '../../domain/read-models/TrainingForAdmin.js';
 import { TrainingSummary } from '../../domain/read-models/TrainingSummary.js';
