+++ api/src/prescription/campaign-participation/infrastructure/repositories/campaign-participant-repository.js
@@ -8,8 +8,8 @@ import {
 import { OrganizationLearnerForStartingParticipation } from '../../../../../lib/domain/read-models/OrganizationLearnerForStartingParticipation.js';
 import { UserIdentity } from '../../../../../lib/domain/read-models/UserIdentity.js';
 import * as campaignRepository from '../../../../../lib/infrastructure/repositories/campaign-repository.js';
-import * as knexUtils from '../../../../../lib/infrastructure/utils/knex-utils.js';
 import { NotFoundError } from '../../../../shared/domain/errors.js';
+import * as knexUtils from '../../../../shared/infrastructure/utils/knex-utils.js';
 import { CampaignParticipant } from '../../domain/models/CampaignParticipant.js';
 import { CampaignToStartParticipation } from '../../domain/models/CampaignToStartParticipation.js';
 import { PreviousCampaignParticipation } from '../../domain/models/PreviousCampaignParticipation.js';
