+++ api/src/prescription/campaign/infrastructure/repositories/campaign-assessment-participation-result-list-repository.js
@@ -2,7 +2,7 @@ import bluebird from 'bluebird';
 
 import { knex } from '../../../../../db/knex-database-connection.js';
 import * as stageCollectionRepository from '../../../../../lib/infrastructure/repositories/user-campaign-results/stage-collection-repository.js';
-import { fetchPage } from '../../../../../lib/infrastructure/utils/knex-utils.js';
+import { fetchPage } from '../../../../shared/infrastructure/utils/knex-utils.js';
 import { CampaignParticipationStatuses } from '../../../shared/domain/constants.js';
 import { CampaignAssessmentParticipationResultMinimal } from '../../domain/read-models/CampaignAssessmentParticipationResultMinimal.js';
