+++ api/src/prescription/campaign/infrastructure/repositories/campaign-participant-activity-repository.js
@@ -1,6 +1,6 @@
 import { knex } from '../../../../../db/knex-database-connection.js';
 import { filterByFullName } from '../../../../../lib/infrastructure/utils/filter-utils.js';
-import { fetchPage } from '../../../../../lib/infrastructure/utils/knex-utils.js';
+import { fetchPage } from '../../../../shared/infrastructure/utils/knex-utils.js';
 import { CampaignParticipationStatuses } from '../../../shared/domain/constants.js';
 import { CampaignParticipantActivity } from '../../domain/read-models/CampaignParticipantActivity.js';
