+++ api/src/prescription/organization-learner/infrastructure/repositories/sco-organization-participant-repository.js
@@ -3,11 +3,11 @@ import _ from 'lodash';
 import { knex } from '../../../../../db/knex-database-connection.js';
 import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../../lib/domain/constants/identity-providers.js';
 import { filterByFullName } from '../../../../../lib/infrastructure/utils/filter-utils.js';
-import { fetchPage } from '../../../../../lib/infrastructure/utils/knex-utils.js';
 import {
   CampaignParticipationStatuses,
   CampaignTypes,
 } from '../../../../../src/prescription/shared/domain/constants.js';
+import { fetchPage } from '../../../../shared/infrastructure/utils/knex-utils.js';
 import { ScoOrganizationParticipant } from '../../domain/read-models/ScoOrganizationParticipant.js';
 
 function _setFilters(qb, { search, divisions, connectionTypes, certificability } = {}) {
