+++ api/src/school/infrastructure/repositories/mission-learner-repository.js
@@ -1,5 +1,5 @@
 import { knex } from '../../../../db/knex-database-connection.js';
-import { fetchPage } from '../../../../lib/infrastructure/utils/knex-utils.js';
+import { fetchPage } from '../../../shared/infrastructure/utils/knex-utils.js';
 import { MissionLearner } from '../../domain/models/MissionLearner.js';
 
 const findPaginatedMissionLearners = async function ({ organizationId, page }) {
