+++ api/src/shared/infrastructure/repositories/user-repository.js
@@ -15,10 +15,10 @@ import { Organization } from '../../../../lib/domain/models/Organization.js';
 import { UserDetailsForAdmin } from '../../../../lib/domain/models/UserDetailsForAdmin.js';
 import { OrganizationLearnerForAdmin } from '../../../../lib/domain/read-models/OrganizationLearnerForAdmin.js';
 import { BookshelfUser } from '../../../../lib/infrastructure/orm-models/User.js';
-import { fetchPage, isUniqConstraintViolated } from '../../../../lib/infrastructure/utils/knex-utils.js';
 import { UserLogin } from '../../../authentication/domain/models/UserLogin.js';
 import { DomainTransaction } from '../../../shared/domain/DomainTransaction.js';
 import { User } from '../../domain/models/User.js';
+import { fetchPage, isUniqConstraintViolated } from '../utils/knex-utils.js';
 
 const getByEmail = async function (email) {
   const foundUser = await knex.from('users').whereRaw('LOWER("email") = ?', email.toLowerCase()).first();
