+++ api/tests/integration/infrastructure/utils/knex-utils_test.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 
 import { knex } from '../../../../lib/infrastructure/bookshelf.js';
-import { DEFAULT_PAGINATION, fetchPage } from '../../../../lib/infrastructure/utils/knex-utils.js';
+import { DEFAULT_PAGINATION, fetchPage } from '../../../../src/shared/infrastructure/utils/knex-utils.js';
 import { databaseBuilder, expect } from '../../../test-helper.js';
 
 describe('Integration | Infrastructure | Utils | Knex utils', function () {
